fix the calculation of getWeekOfDate
The previous calculation didn't properly handle the case where a week
spanned from the end of one year to the start of the next. Replace it
with a simpler approach that just calculates the number of weeks since
the epoch, shifted so weeks start on Monday.

Then use this week number when generating the times for the weekly data
so that it doesn't end up with a partial week for the end of the time
period.